src/main/webapp/app/entities/address/address.service.ts
import axios from 'axios';

import { IAddress } from '@/shared/model/address.model';

const baseApiUrl = 'api/addresses';

export default class AddressService {
  public find(id: number): Promise<IAddress> {
    return new Promise<IAddress>(resolve => {
      axios.get(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public retrieve(): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(baseApiUrl).then(function(res) {
        resolve(res);
      });
    });
  }

  public delete(id: number): Promise<any> {
    return new Promise<any>(resolve => {
      axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res);
      });
    });
  }

  public create(entity: IAddress): Promise<IAddress> {
    return new Promise<IAddress>(resolve => {
      axios.post(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public update(entity: IAddress): Promise<IAddress> {
    return new Promise<IAddress>(resolve => {
      axios.put(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }
}
